requirements.txt
# Core dependencies
python>=3.10
GitPython>=3.1.40
tree-sitter>=0.20.4
tree-sitter-python>=0.20.4
tree-sitter-java>=0.20.4
pydantic>=2.5.0
pyyaml>=6.0.1

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Utilities
tqdm>=4.66.1
python-dateutil>=2.8.2
